requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argparse==1.4.0           # via dateutils
backports-abc==0.5        # via tornado
backports.shutil-get-terminal-size==1.0.0  # via ipython
beautifulsoup4==4.6.0     # via django-htmlmin
beautifulsoup==3.2.1
bleach==2.1.3             # via nbconvert
certifi==2018.4.16        # via requests, tornado
chardet==3.0.4            # via requests
configparser==3.5.0       # via entrypoints
cssselect==1.0.3          # via premailer
cssutils==1.0.2           # via premailer
dateutils==0.6.6
decorator==4.3.0          # via ipython, traitlets
django-bootstrap-breadcrumbs==0.9.1
django-colorfield==0.1.14
django-crispy-forms==1.6.0
django-debug-toolbar==1.3.2
django-extensions==1.6.7
django-filemanager==0.0.5
django-filter==0.10.0
django-haystack==2.4.1
django-htmlmin==0.10.0
django-jfu==1.0.7
django-recaptcha==1.4.0
django-reversion==1.8
django==1.5.12
django_premailer==0.2.0
entrypoints==0.2.3        # via nbconvert
enum34==1.1.6             # via traitlets
flup==1.0.3
functools32==3.2.3.post2  # via jsonschema
html5lib==1.0.1           # via bleach, django-htmlmin
idna==2.6                 # via requests
ipdb==0.10.2
ipykernel==4.5.2
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython[notebook]==5.3.0
ipywidgets==7.2.1         # via ipython
jinja2==2.10              # via nbconvert, notebook
jsonschema==2.6.0         # via nbformat
jupyter-client==5.2.3     # via ipykernel, notebook
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook
lxml==4.2.3               # via premailer
markupsafe==1.0           # via jinja2
mistune==0.8.3            # via nbconvert
nbconvert==5.3.1          # via notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.5.0           # via ipython, widgetsnbextension
pandocfilters==1.4.2      # via nbconvert
pathlib2==2.3.2           # via ipython, pickleshare
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.2.0
premailer==3.0.0          # via django-premailer
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.5
ptyprocess==0.6.0         # via pexpect, terminado
pycrypto==2.6.1
pygments==2.2.0           # via ipython, nbconvert
python-dateutil==2.7.3    # via dateutils, jupyter-client
python-gettext==3.0
python-gitlab==1.6.0
python-magic==0.4.15
pytz==2018.5              # via dateutils
pyzmq==17.0.0             # via jupyter-client, notebook
requests==2.18.4
scandir==1.7              # via pathlib2
send2trash==1.5.0         # via notebook
simplegeneric==0.8.1      # via ipython
singledispatch==3.4.0.3   # via tornado
six==1.11.0               # via bleach, django-bootstrap-breadcrumbs, django-extensions, html5lib, pathlib2, prompt-toolkit, python-dateutil, python-gitlab, singledispatch, traitlets, transliterate
south==1.0.2
sqlparse==0.1.19
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
tornado==4.5.3
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
transliterate==1.10.2
unidecode==1.0.22
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
werkzeug==0.14.1
whoosh==2.7.4
widgetsnbextension==3.2.1  # via ipywidgets
xmltodict==0.11.0
yandex-oauth-py==0.1.1
